// src/types/Effect.ts
export type EffectId =
  | "CLEAR_BUILDING_FARM_GROWTH"
  | "CLEAR_UI_ABSORBER_CHARGE"
  | "CLEAR_UI_OVERHEATING_CRITICAL"
  | "CLEAR_UI_OVERHEATING_HOT"
  | "CLEAR_UI_OVERHEATING_VERY_HOT"
  | "CLEAR_UI_WINDOW"
  | "DESTROY"
  | "ON_COLONIST_DESTROYED"
  | "ON_FARM_WORKED"
  | "ON_ROAD_BUILD"
  | "RESET_WORK_CONTRIBUTED"
  | "SHIELD_DISCHARGE"
  | "SPAWN_BUILDING_WALL_CRACKED"
  | "SPAWN_BUILDING_WALL_CRUMBLING"
  | "SPAWN_ENEMY_DRONE"
  | "SPAWN_PLAYER_CORPSE"
  | "SPAWN_UI_OVERHEATING_CRITICAL"
  | "SPAWN_UI_OVERHEATING_HOT"
  | "SPAWN_UI_OVERHEATING_VERY_HOT";

export interface AllEffect {
  ALL: Effect[];
}

export type Effect = EffectId | AllEffect;